Move listen loading to Sass::Util so Compass can use it.
